Move the Submission object to the Submission domain
