+++ src/App/Bundle/MainBundle/Infrastructure/Submission/DoctrineWriter.php
@@ -3,7 +3,7 @@
 namespace App\Bundle\MainBundle\Infrastructure\Submission;
 
 use App\Bundle\MainBundle\Entity\Submission as SubmissionEntity;
-use App\Domain\Submission;
+use App\Domain\Submission\Submission;
 use App\Domain\Submission\WriterInterface;
 use Doctrine\ORM\EntityManagerInterface;
