+++ src/App/Domain/Submission/Submission.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Domain;
+namespace App\Domain\Submission;
 
 /**
  * A post submission
